Add files via upload
This is my first capstone project for the Hyperion Dev Software Engineering Bootcamp.
The project scope included creating two calculators, one to output the earnings on an investment and the other to calculate the monthly payments on a home loan. The instructions were to use if-elif-else control instructions based on user input and also use defensive programming to foresee any possible errors.
I'm working on improving the defensive programming and I also want to add a currency selector and decimal separators for the numbers the user is seeing on the screen.